# swagger.yaml
host: "localhost:8080"
schemes:
    - http
securityDefinitions:
    Bearer:
        type: apiKey
        name: Authorization
        in: header
        description: enter the token with the Bearer prefix e.g. Bearer eyJhbGci...
definitions:
    user:
        properties:
            created_at:
                description: The timestamp when the user was created.
                example: '"2022-03-17T09:00:00Z"'
                format: date-time
                type: string
            password_changed_at:
                description: The timestamp when the password was last changed.
                example: '"2022-03-17T10:00:00Z"'
                format: date-time
                type: string
            role:
                description: The role of the user.
                example: administrator
                type: string
            username:
                description: The username of the user.
                example: vk-user
                type: string
        title: userResponse represents the response body for a user.
        type: object
...
    userRequest:
        type: object
        required: 
            - username
            - password
        properties:
            username:
                description: The username of the user.
                example: vk-user
                type: string
            password:
                description: The password of the user.
                example: password
                type: string
...
info: {}
paths:
    /users:
        post:

            operationId: createUser
            consumes:
                - application/json
            operationId: createUser
            summary: Creates a new user.
            parameters:
                - in: body
                  name: user
                  description: The user to create.
                  schema:
                    $ref: '#/definitions/userRequest'
            produces:
                - application/json
            tags:
                - users
            responses:
                200:
                    $ref: '#/responses/userResponse'
                400:
                    $ref: '#/responses/error400Response'
                401:
                    $ref: '#/responses/error401Response'
                403:
                    $ref: '#/responses/error403Response'
                500:
                    $ref: '#/responses/error500Response'
    /users/login:
        post:
            operationId: loginUser
            summary: Logs in a user.
            parameters:
                - in: body
                  name: user
                  description: The user to log in.
                  schema:
                    $ref: '#/definitions/userRequest'
            produces:
                - application/json
            responses:
                200:
                    $ref: '#/responses/loginUser'
                400:
                    $ref: '#/responses/error400Response'
                401:
                    $ref: '#/responses/error401Response'
                500:
                    $ref: '#/responses/error500Response'
            tags:
                - users
...
responses:
...
    loginUser:
        description: loginUserResponse represents the response body for logging in a user.
        headers:
            access_token:
                description: The access token for the user.
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6Im...
                type: string
        properties:
            user:
                description: The user information.
        schema:
            $ref: '#/definitions/user'
...
    userResponse:
        type: object
        description: userResponse represents the response body for a user.
        properties:
            user:
                description: User Information.
        schema:
            $ref: '#/definitions/user'
    error400Response:
        description: Bad Request.
    error401Response:
        description: Unauthorized.
    error403Response:
        description: Forbidden.
...
    error500Response:
        description: Internal Server Error.
swagger: "2.0"